Porting shufflenetv2
